fix(project): Integrate VSCode with linting/formatting
I've moved config files to the root as that is the default place most
tools expect their configs, it save extra customization in usage, so
makes command simpler. Adding the `.vscode` directory with settings
for VSCode means that the linting/formatting tools are integrated with
the IDE and are applied when a file is saved, as opposed to at commit
time.

I've also made the absolute imports, which was setup, but never really
used, so we can get rid of lots of confusing import statments like

`../../../blah`

Figuring out that directory gets harder the deeper the import hierarchy
gets. Using absolute paths means files can move around without every
having to change, unlike the relative imports used here.

IssueID NCEA-22