fix(project): Update and refactor npm scripts
I've renamed a couple of scripts to simplify and fixed the linting
one that seemed to want to find/lint sass files, by not including
`--ext` flag. I also ran `npm run build` that may have build the files
and therefore fixed the issue as well, it's not exactly clear.