fix(project): Add asdf version manager config
It's nice that there is an `engines` block in the package.json file
but it states anything greater than node 16, which is ancient, yet
the Dockerfile is using the latest node alpine which is version 20.x.
I have added this file as it configures my system to use the appropriate
Node version for this project as I have many projects using different
versions. Devs may use other such tools e.g. nvm and should also add
those, if they like.

IssueID #NCEA-22