fix(project): Add a docker compose file to make running app locally easier
The app runs inside a docker container when deployed, but unless we
manually build and run the container locally, we don't really know
if it works, until it is deployed. Instead of dealing with more complex
Docker commands, docker compose makes it trivial to build and run the
container locally, while developing.

IssueID #NCEA-22