feat(filters): Implement filter functionality
The new DSP filters now are functional and make a new GET request whenever one is changed (the data is still mocked and the filters do not change the returned data).
The option to change data scope is functional, and will show/hide other filters depending on whether it is selected.

There still exist some issues to cleared up. There are a few UX issues like text inputs always refreshing when they lose focus and accordions closing on refresh.
How the filters will interact with the new API is also unclear, which means it is currently still possible to get non-NCEA results when the data scope is limited to NCEA.

IssueID #NCEA-123