fix(filters): Prevent invalid state in url
As the form is submitted every time a filter is changed, the url must represent a valid state at all times, in order to get correct data.
It was previously possible to select 2 organisations, one with NCEA data and one without, then change the data scope which would cause the url to become invalid, since it included an organisation that should not be there.
This fix first validates the form data and then constructs the query paramaters to make sure the url is always valid.

IssueID #NCEA-123