fix(filters): Fix styling issue in classifier search
A previous commit to format code revealed an issue in the
`application.css` where a media query was disabled by appending
`\0` to the front of the condition. This did not error in the
minified code, but upon formating it caused an error.
I removed the `\0` thinking it was a mistake, but it was purposefuly
added to disable those styles, so I removed them in this change.