fix(filters): Fix query paramters removed from url upon filter change
The paramaters in the url that describe things like the query,
results per page, sort, etc, were being replaced when the filters
were updated the and the url was changed.
This fix copies those query params into the new url so all meta
parameters are restored when a filter is changed.

IssueID #NCEA-123