feat(filters): Implement filter behaviour
Previously, the new filters did not apply to the mock data. This
can make it a bit more difficult to do testing, so this commit
implements filter behaviour on the mock data with the selected
filters so it is more accurate to when the new API is put in place.

IssueID #NCEA-138