feat(frontend): Add login/logout functionality
I have been investigating how hapi.js manages cookies and how to
read and parse a cookie to extract user info, to then make it
available on each `route`. I have accomplished this, just not in
the way the docs state as I couldn't get it from the `request.state`,
so I read the cookies directly and manually extract the JWT and
add that to the `request` object, to be used by the route controller.

I have added some UI bit *hard-coded) for now, to state if a user is
logged in and to display the appropriate login/logout link in each
circumstance.

IssueID NCEA-135