fix(filters): Move styles to SCSS
`application.css` is the compiled CSS that comes from all the SCSS
files, so making any modifications in that file means they will be
removed eventually.
I have moved all of the new filter styles into the respective SCSS
file so it is correctly compiled.
By making this change, it also broke Mark's change to the base paths.
Luckily, GOVUK provides a SCSS variable that defines the base path
for the assets, so by changing that it once again points the font
to the correct place.