feat(filters): Prevent map from disappearing due to lack of spatial data
With the new filters, an option has been added to filter all results
without spatial data. This means there would be nothing available
to display on the map, so it got hidden but that is very jarring
to the user and makes for a worse user experience.
Instead, the map is always shown but the background is made
grayscale and a message is shown indicating the lack of spatial
data.
This prevents the content jumping around and is much nicer to use.

IssueID #NCEA-5