chore(filters): Consolidate date functions to single fie
There was already a file dedicated to formatting dates, so I
renamed the file to `dates.ts` to be be more generic, and moved
the date function used in the filtering to the new dates file.

This does include a lot of formatting as the find/replace meant
the files were formatted.