fix(auth): Alter auth implemenation to use credentials directly
The function that returned user details from given user data
has been instead replaced with just accessing the data directly.
If there data is `null` the user is logged out, and if it's not
null then we can access the email directy from the data.

This also moves the auth logic out of an event handler in the server,
to a dedicated custom auth scheme in another file, which I feel
makes the server initialisation code neater, and keeps code with
different functionality more separated.

IssueID #NCEA-135