fix(tests): Correct broken auth and server tests
Changing the auth implementation to a schema broke the server tests
so I corrected those.
I also updated the tests in `HomeController` as they weren't really
testing the JWT side of authentication. Now, a mock JWT is passed
into the schema and the decoded data is passed to the controller.
This makes sure the auth schema does not break in the future.

IssueID #NCEA-135