feat(auth): Implement logout system
It was recommended that we handle logging out ourselves, so this
introduces that feature. A new `/logout` endpoint has bee added
and the logout button makes a `GET` request to this, including a
`redirect_uri` query parameter which will redirect the user back
to the page they logged out from.
It removes the cookie from the browser so they appear logged out.

This also fixes any respective tests.

IssueID #NCEA-135