WIP feat(search): Replace mocked search data with new API
Now that we have a queryable search API, we can replace the mock
data. The API doesn't return data exactly how we want and it will
change, so we will have to revisit the transformation function(s)
that take the API response and re-shape the data to how we want it.