feat(filters): Implement map filters
This commit adds the new filters into the map modal and makes them
functional. The behaviour is slightly different as the webpage
should not reload every time you change a filter, othwerwise it
would kick you out of the modal which is bad UX,  and instead it
should fetch and update the results in the background.
The filters are now saved when you exit the modal, so any filters
applied on the map stay applied when exiting the modal.

This also fixes any respective tests.

IssueID #NCEA-5