fix(filters): Fix edge case where map disappears with lack of spatial data
A previous commit changed the behaviour of the map so if a query
returned no results, rather than the map disappearing, which is
jarring to the user, the map stays but a message is shown saying
there is a lack of spatial. This was implemented by checking if
the number of results returned was equal to 0, but that did not
account for the case where a user had applied the `Non-spatial`
data only filter. This would return >0 results but none with
spatial data so the map would disappear.

IssueID #NCEA-5