feat(filters): Add clear/apply filter buttons to results view
The form previously automatically submitted for every filter change.
This lead to poor UX if the user wanted to select multiple filters.
Now, there are 2 buttons which either clear or apply the selected
filters, leaving the choice up to the user instead of the site
guessing on their behalf.

IssueID #NCEA-172