feat(frontend): Add sytem for including svg icons
There was no system that allowed for SVG icons to be reused which
has led to much code duplication in the nunjucks files. I have written
a small system that can be easily extended to allow SVG icons to be
reused in the code without having to copy and paste or make it a
static asset. It also allows for the attributes of the SVG to be
extended which means things like width, height and strokes can
easily be changed.