feat(badges): Implement badges and badge styling
When a filter is applied a badge should be shown at the top of the
results so a user can clearly see the currently active filters.
This is especially important as the accordions close so they wont
be able to see the selected filters once the page has loaded.

Currently, the badges are added to the DOM and they have the same
styling as per DSP, but they do not yet function when you lick on
them. When clicked, they should remove the respective filter that
matches the badge that was clicked on.

IssueID #NCEA-173