chore(filters): Reduce number of magic strings
A lot of the filter IDs used hardcoded strings that were referenced
in many different files. There is an object that already exists,
that is passed into the view that contains different IDs for each
filter. I have replaced the magic strings with the strings from
within the object so all IDs are generated from a single location,
unifying the scripts a little bit more and making it easier to
add and change filter names in the future.

Unfortunately this doesn't stop the fact that the format of the
IDs is duplicated in the nunjucks file and in the JS. It could be
beneficial to make a nunjucks macro to generate the filter so
that the ID format is always the same for every filter without
the possibility of another filter being added with a different ID
format. The macro could potentially create some JS that also
injects the IDs into a global object so that the scripts do not
have to write them out manually as well.

IssueID #NCEA-173