fix(badges): Remove interactivity from badges
We have a agreed a new system for the badges which changes their
current behaviour. Interactive badges will only show under the
keyword filter, so users can see what keywords they intend to
apply, and have the ability to remove them.
The badges that show above the search results are now informational
only and are there to inform the user about their currently applied
filter, but they cannot click on the badge to remove it.
This is to keep the behaviour of the filters consistent and more
sensible with the `Apply Filters` button.

IssueID #NCEA-173