feat(filters): Make `All` checkbox visually select/deselect all
The `All` checkbox used to be handled entirely in logic, and so
if you selected it, it would not  not visually select every
other checkbox in that category.
Due to the change in badge behaviour, there needed to be a more
flexible way to deselect filters from a single category, rather
than clearing every single filters. By making `All` select all,
the user can check it, to select everything, then uncheck it to
deselect everything, which will clear the selected filters in that
category.

IssueID #NCEA-183