fix(search): Update mapping between API response and UI data shape
Again, since we are moving from data provided by ElasticSearch to
data provided by an AGM API, the response from the API differs greatly.
I have written a new `transformSearchResponse` function that takes
the API response and generates the data shape the app was already
using. This is greately simplified over what was retrieved from
ElasticSearch.

NOTE: I have left the old `formatSearchResponse` function as it is
still used by the `getDocumentDetails`. That will be updated in
future commits though.