feat(filters): Make scope buttons refresh automatically
When the filters were updated to use buttons to submit instead of
submitting automatically, that also meant the scope buttons did not.
We discovered this wasnt very user friendly as they could pointlessly
select filters that would be removed once submitting.

We decided that the scope buttons should not be classed as filters
and should have their own heading. This means they should also
have different behaviour to the filters which allows them to
automatically apply when changed.
If you have filters select that are not applied when you change scope,
those filters will be applied.

The map view now also has the 2 different headings for the dat
scope and filters.

IssueID #NCEA-182