fix(filters): Update filter interfaces to use union for scope instead of boolean
The original filter interface stored an optional boolean that
represented whether the filter was part of NCEA metadata scope
only. This made lots of logic convoluted and difficult to explain
to others, which will be required for future changes. It also
meant no other scope could be added.
I have since updated it so each filter stores a union of enum
values which represent each scope. This simplifies the logic a lot
and makes adding scopes in the future more simple.

IssueID #NCEA-174