fix(filters): Update data format filter values
The search API expects an exact match for the data format filter.
I originally use makeshift values for each filter which was just
the file extension but this means the API will never return results
for any format as it doesnt match exactly whats stored. I have instead
pulled the values from DSP and used those which should match what is
stored on each record in the database so the filter will work.

IssueID #NCEA-184