feat(map): Add clustering to map and viewable list of clustered records
We discovered that the data returned by the new seach API often had
tens or hundreds of markers sitting directly on top of each other
in the map. This can be confusing if the map says to view "X" number
of records, but it looks like half are missing.
It was decided that `clustering` was the solution to this, so any
close markers would be grouped into a cluster and a number would
display how many records are in that cluster.

The cluster also needed to be interactable, similarly to how a single
marker was before, but instead it should show a list of every record
contained in that cluster, and clicking on any of those records
should open the normal small info popup.
Any single markers not part of a cluster have not had any changes.

IssueID #NCEA-185