fix(elements): Filtering for all on org or Data Format is unchecking some checkboxes
 Issue: There is a function called "readQueryParams" from this we are reading the filter values and passed it to the result API, this "readQueryParams"
 function will return the query string with comma separated. when it comes to organization or Data format, in which  some filter items which contain in name it self a comma character is there, hence it treated as a separate filter value when it doing the query string split by comma. this comma separated values contains some other  name which are not expected values.

Fix: I tried to modify "readQueryParams"  function it will return the array of filter items when key value  is available, but it is used in so many places in the source code, for avoiding complexity and more clarity, I created the new function  called " readListQueryParams ", it will only get the list items format from the query params in array format, which are required for the org, data formats, keywords which are passed to the result API. It also avoids the spiting of  the query string items in other places like keywords, data formate, because we spitted at this function level.

Issued Id: NCEA-231.