fix(frontend): Display "Unavailable" in place of null dates
Sometimes the study period has dates that are `null` in the data.
Because of behaviour in `date-fns`, these were being formatted
as `1st January 1970`, which in this case is confusing and doesnt
make sense. Instead, it should show "Unavailable" in place of these
missing dates, which is the behaviour I have added.

IssueID #NCEA-230