fix(filters): Integrate the sort by UI widget
Until now we have ignored the sorting of data results as the API
didn't have the ability. Now it does, so Ihave added it to the
URL as a query param. I would have preferred this filter be like all
the rest and be part of the POST body, but this is how AGM have
coded the API.