fix(layout): Update link for NCEA logo
The link was only pointing to the landing page, meaning we had 3
links there pointing to the same place, so updated it with the site
recommended by the client.

I also noticed the link didn't open in a new tab as it was supposed
to, so fixed that at the same time.