feat(search): Cache API results
I have added code to Cache API responses per user, to reduce calls
to the API. The key of the cache is based on the querystring and
filters set, so if any of those change, a new request is made. I
omit the pagination `pg` and `resultsPerPage` values as they change
due to pagination, not due to results changing.