fix(keyboard-filters): Restore client-side keywords functionality when scope is changed
Previously, switching the scope from `scope=all` to `scope=ncea` triggered a DOM refresh to update map results, but the client-side code handling keywords functionality was lost during this refresh. This commit reintroduces the keywords functionality logic so that it persists and works correctly after the scope change.

IssueID NCEA-297