fix(moreinfo): Update the resources name when it doesn't exist
It is possible that a record has links to resources, but the metadata
isn't great, i.e. missing the name. In this case we were displaying
`N/A`, but it is preferred we use the same text as the DSP Platform,
so I have changed it to `Download data`, only if there is no name.